pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "mansur", email = "mansurmekin@gmail.com"},
]
dependencies = [
    "django==4.2",
    "django-filter>=23.5",
    "djangorestframework>=3.14.0",
    "black>=23.11.0",
    "python-dotenv>=1.0.0",
    "drf-yasg>=1.21.7",
    "setuptools>=69.0.2",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

